feat : 채널별 판매중 상품 목록 매퍼 구현

+++ src/main/java/susussg/pengreenlive/broadcast/dto/BroadcastDTO.java
@@ -14,7 +14,7 @@ import java.util.Date;
 @NoArgsConstructor
 public class BroadcastDTO {
     //방송
-    private int broadcastSeq;
+    private long broadcastSeq;
     private String channelNm;
     private String broadcastTitle;
     private byte[] broadcastImage;
